requirements.txt
dill==0.3.3
Flask==2.0.1
grpcio==1.36.1
grpcio-tools==1.36.1
redis==3.5.3
requests==2.25.1
